requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
aiohttp==3.10.5
beautifulsoup4==4.12.3
feedparser==6.0.11
google-genai>=0.8.0
python-dotenv==1.0.1
lxml==5.3.0
nest-asyncio==1.6.0
pyngrok==7.2.0
requests==2.32.3
